simple multi timeframe tops moved timeframe to 1st index (index[0])

+++ Config.py
@@ -1,6 +1,7 @@
 import base64
 import hashlib
 import json
+import os
 from datetime import timedelta
 from enum import Enum
 from typing import List
@@ -47,7 +48,11 @@ class Config():
         self.multi_timeframe_ohlc_columns = self.ohlc_columns + ['timeframe']
         self.multi_timeframe_ohlca_columns = self.ohlca_columns + ['timeframe']
         self.multi_timeframe_peaks_n_valleys_columns = self.ohlc_columns + ['timeframe', 'peak_or_valley'] # 'strength',
-        self.multi_timeframe_trend_boundaries_columns = ['timeframe', 'end', 'bull_bear_side']
+        self.multi_timeframe_trend_boundaries_columns = ['timeframe', 'end', 'bull_bear_side',
+                                                         'highest_hig', 'lowest_low', 'high_time', 'low_time',
+                                                         'trend_line_acceleration', 'trend_line_base',
+                                                         'canal_line_acceleration', 'canal_line_base',
+                                                         ]
 
         self.end_time = '2021-03-01 03:43:00'
 
@@ -73,8 +78,9 @@ config_digest = str.translate(base64.b64encode(hashlib.md5(config_as_json.encode
                               .decode('ascii'), {ord('+'): '', ord('/'): '', ord('='): '', })
 
 if DEBUG: print(config_digest)
-with open(f'Config.{config_digest}.json', 'w+') as config_file:
-    config_file.write(str(config_as_json))
+if os.path.exists(f'Config.{config_digest}.json'):
+    with open(f'Config.{config_digest}.json', 'w+') as config_file:
+        config_file.write(str(config_as_json))
 
 config.id = config_digest
